+++ CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 This project uses [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Added
+
+- Added the `async-tempfile` crate feature to enable the `SharedTempFile`
+  type.
+- Added the generic `SharedFile<T>` type for use with arbitrary `AsyncWrite + AsyncRead` types.
+
 ## [0.0.1] - 2023-06-13
 
 ### Internal